better grouped table interface with accordions
